pyproject.toml
[tool.poetry]
name = "blog-service"
version = "0.1.0"
description = ""
authors = ["locmai <locmai0201@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.4"
fastapi = "^0.53.1"
dnspython = "^1.16.0"
motor = "^2.1.0"
uvicorn = "^0.11.3"
loguru = "^0.4.1"
pyjwt = "^1.7.1"
bcrypt = "^3.1.7"
passlib = "^1.7.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
autopep8 = "^1.5"
pytest = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
